[PON-42] Change Random Wheel color
